requirements.txt
fastapi
uvicorn
pydantic
redis>=5.0.0
python-dotenv>=1.0.1
groq>=0.11.0
sqlalchemy>=2.0.0
asyncpg